+++ src/lib/ui/SvastView.svelte
@@ -1,6 +1,4 @@
 <script lang="ts">
-	import {EMPTY_OBJECT} from '@feltjs/util/object.js';
-
 	import {getApp} from '$lib/ui/app';
 	import {type ViewNode, toViewProps} from '$lib/vocab/view/view';
 
@@ -10,7 +8,7 @@
 
 	export let view: ViewNode;
 
-	$: props = toViewProps(view) || EMPTY_OBJECT;
+	$: props = toViewProps(view);
 </script>
 
 <!-- TODO can we format this without adding unwanted whitespace? -->